Fix alertMessage type handling
* Alert reducer was written expecting strings, but all the catch
dispatches were feeding it an Error object. This adds some type handling
so that it will accept either a string, or an object. This isn't what we
should accept in the long term, so we need to file an issue to capture
this and hopefully introduce some stronger compile time checks.