Early discovery service cert acceptance
It's been agreed that it's a better experience for users to accept the
discovery service cert as part of app initialization with the rest of
the certificates that have to be accepted instead of having it interrupt
the middle of plan creation.

Because the debug tree feature also depends on this cert acceptance,
it's a good time to bump the initial call to app initialization. This PR
introduces the check during app init, and makes sure to wait until the
actual host MigCluster has been loaded, since it's actually possible the
name of that cluster will differ from the default value.