Behave the same as 1.8.4 on a new migration plan (#1526)
The loading behavior was modified in 1.8.5 and
the grid loading was not behaving the same as
1.8.4 when creating a new direct volume migration
plan. The spinner was not covering the grid and
the grid was showing the empty state.

Instead revert to covering the entire grid. And
only show the validating spinner when the grid
selection changed.

Signed-off-by: Alexander Wels <awels@redhat.com>